+++ tests/testthat/test-half.life.R
@@ -323,19 +323,23 @@ test_that("for an interval with multiple doses lambda.z is calculated for the la
   expect_equal(lzs[1], lzs[3])
 })
 
-test_that("get_halflife_points", {
+test_that("get_halflife_points method (1: PKNCAdata, 2: PKNCAresults)", {
+  # Standard analysis ----
   o_conc <- PKNCAconc(Theoph, conc~Time|Subject)
   o_data <- PKNCAdata(o_conc, intervals = data.frame(start = 0, end = Inf, half.life = TRUE))
   o_nca <- suppressMessages(pk.nca(o_data))
-  hl_points <- suppressMessages(get_halflife_points(o_nca))
+
+  hl_points1 <- suppressMessages(get_halflife_points(o_data))
+  hl_points2 <- suppressMessages(get_halflife_points(o_nca))
 
   # Visualize the results
   # d_all <- Theoph
-  # d_all$hl_points <- hl_points
-  # ggplot(d_all, aes(x = Time, y = conc, colour = hl_points, groups = Subject)) +
+  # d_all$hl_points2 <- hl_points2
+  # ggplot(d_all, aes(x = Time, y = conc, colour = hl_points2, groups = Subject)) +
   #   geom_point() + geom_line()
+  expect_equal(hl_points1, hl_points2)
   expect_equal(
-    hl_points,
+    hl_points2,
     c(FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, TRUE,
       TRUE, TRUE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE,
       TRUE, TRUE, TRUE, TRUE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE,
@@ -352,60 +356,111 @@ test_that("get_halflife_points", {
       FALSE, FALSE, FALSE, TRUE, TRUE, TRUE)
   )
 
-  # start != 0 (#470)
+  # Other analyses that are dependent on half-life trigger the half-life calculation ----
+  o_conc <- PKNCAconc(Theoph, conc~Time|Subject)
+  o_data <- PKNCAdata(o_conc, intervals = data.frame(start = 0, end = Inf, aucinf.obs = TRUE))
+  o_nca <- suppressMessages(pk.nca(o_data))
+
+  hl_points1 <- suppressMessages(get_halflife_points(o_data))
+  hl_points2 <- suppressMessages(get_halflife_points(o_nca))
+
+  # Visualize the results
+  # d_all <- Theoph
+  # d_all$hl_points2 <- hl_points2
+  # ggplot(d_all, aes(x = Time, y = conc, colour = hl_points2, groups = Subject)) +
+  #   geom_point() + geom_line()
+  expect_equal(hl_points1, hl_points2)
+  expect_equal(
+    hl_points2,
+    c(FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, TRUE,
+      TRUE, TRUE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE,
+      TRUE, TRUE, TRUE, TRUE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE,
+      FALSE, FALSE, TRUE, TRUE, TRUE, FALSE, FALSE, FALSE, FALSE, FALSE,
+      FALSE, FALSE, FALSE, TRUE, TRUE, TRUE, FALSE, FALSE, FALSE, FALSE,
+      FALSE, FALSE, FALSE, TRUE, TRUE, TRUE, TRUE, FALSE, FALSE, FALSE,
+      FALSE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, FALSE, FALSE,
+      FALSE, FALSE, FALSE, FALSE, FALSE, TRUE, TRUE, TRUE, TRUE, FALSE,
+      FALSE, FALSE, FALSE, FALSE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE,
+      FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, TRUE,
+      TRUE, TRUE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE,
+      FALSE, TRUE, TRUE, TRUE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE,
+      FALSE, FALSE, TRUE, TRUE, TRUE, FALSE, FALSE, FALSE, FALSE, FALSE,
+      FALSE, FALSE, FALSE, TRUE, TRUE, TRUE)
+  )
+
+
+  # start != 0 (#470) ----
   o_data_nzstart <- PKNCAdata(o_conc, intervals = data.frame(start = 5, end = Inf, half.life = TRUE))
   o_nca_nzstart <- suppressMessages(pk.nca(o_data_nzstart))
-  hl_points_nzstart <- suppressMessages(get_halflife_points(o_nca_nzstart))
-  # Find the specific rows that have differences
-  expect_equal(which(!is.na(hl_points_nzstart) & hl_points != hl_points_nzstart), c(62, 63, 84))
-  expect_true(all(is.na(hl_points_nzstart[Theoph$Time < 5])))
-  expect_true(!any(is.na(hl_points_nzstart[Theoph$Time > 5])))
 
-  # Setup for the remaining tests
+  hl_points1_nzstart <- suppressMessages(get_halflife_points(o_data_nzstart))
+  hl_points2_nzstart <- suppressMessages(get_halflife_points(o_nca_nzstart))
+  # Find the specific rows that have differences
+  expect_equal(hl_points1_nzstart, hl_points2_nzstart)
+  expect_equal(which(!is.na(hl_points2_nzstart) & hl_points2 != hl_points2_nzstart), c(62, 63, 84))
+  expect_true(all(is.na(hl_points2_nzstart[Theoph$Time < 5])))
+  expect_true(!any(is.na(hl_points2_nzstart[Theoph$Time > 5])))
+
+  # Setup for the remaining tests ----
   d_conc <- as.data.frame(Theoph[Theoph$Subject %in% Theoph$Subject[1], ])
   d_conc$incl <- c(FALSE, TRUE, rep(NA, 8), TRUE)
   d_conc$excl_txt <- c(NA, "x", rep(NA, 8), "x")
   d_conc$excl <- c(NA, TRUE, rep(NA, 6), TRUE, FALSE, TRUE)
 
-  # No modification
+  ## No modification/exclusion ----
   o_conc <- PKNCAconc(d_conc, conc~Time|Subject)
   o_data <- PKNCAdata(o_conc, intervals = data.frame(start = 0, end = Inf, half.life = TRUE))
   o_nca <- suppressMessages(suppressWarnings(pk.nca(o_data)))
+  hl_points1 <- suppressMessages(get_halflife_points(o_data))
+  hl_points2 <- suppressMessages(get_halflife_points(o_nca))
+  expect_equal(hl_points1, hl_points2)
   expect_equal(
-    suppressMessages(get_halflife_points(o_nca)),
+    hl_points1,
     c(rep(FALSE, 8), rep(TRUE, 3))
   )
 
-  # Test include_half.life
+  ## Test include_half.life ----
   o_conc <- PKNCAconc(d_conc, conc~Time|Subject, include_half.life = "incl")
   o_data <- PKNCAdata(o_conc, intervals = data.frame(start = 0, end = Inf, half.life = TRUE))
   o_nca <- suppressMessages(suppressWarnings(pk.nca(o_data)))
+
+  hl_points1 <- suppressMessages(suppressWarnings(get_halflife_points(o_data)))
+  hl_points2 <- suppressMessages(get_halflife_points(o_nca))
+  expect_equal(hl_points1, hl_points2)
   expect_equal(
-    suppressMessages(get_halflife_points(o_nca)),
+    hl_points1,
     c(FALSE, TRUE, rep(FALSE, 8), TRUE)
   )
 
-  # Test exclude
+  ## Test exclude ----
   o_conc <- PKNCAconc(d_conc, conc~Time|Subject, exclude = "excl_txt")
   o_data <- PKNCAdata(o_conc, intervals = data.frame(start = 0, end = Inf, half.life = TRUE))
   o_nca <- suppressMessages(pk.nca(o_data))
+
+  hl_points1 <- suppressMessages(get_halflife_points(o_data))
+  hl_points2 <- suppressMessages(get_halflife_points(o_nca))
+  expect_equal(hl_points1, hl_points2)
   expect_equal(
-    suppressMessages(get_halflife_points(o_nca)),
+    hl_points2,
     # NA values indicate that the point was excluded from all calculations
     c(FALSE, NA, FALSE, FALSE, FALSE, FALSE, FALSE, TRUE, TRUE, TRUE, NA)
   )
 
-  # Test exclude_half.life
+  ## Test exclude_half.life ----
   o_conc <- PKNCAconc(d_conc, conc~Time|Subject, exclude_half.life = "excl")
   o_data <- PKNCAdata(o_conc, intervals = data.frame(start = 0, end = Inf, half.life = TRUE))
   o_nca <- suppressMessages(pk.nca(o_data))
+
+  hl_points1 <- suppressMessages(get_halflife_points(o_data))
+  hl_points2 <- suppressMessages(get_halflife_points(o_nca))
+  expect_equal(hl_points1, hl_points2)
   expect_equal(
-    suppressMessages(get_halflife_points(o_nca)),
+    hl_points1,
     # NA values indicate that the point was excluded from all calculations
     c(FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, TRUE, TRUE, FALSE, TRUE, FALSE)
   )
 
-  # Multiple, overlapping half-life calculations gives an error
+  ## Multiple, overlapping half-life calculations gives an error ----
   o_conc <- PKNCAconc(d_conc, conc~Time|Subject)
   o_data <-
     PKNCAdata(
@@ -417,12 +472,12 @@ test_that("get_halflife_points", {
         )
     )
   o_nca <- suppressMessages(suppressWarnings(pk.nca(o_data)))
-  expect_error(
-    suppressMessages(get_halflife_points(o_nca)),
-    regexp = "More than one half-life calculation was attempted on the following rows: 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11"
-  )
 
-  # Half-life points are selected right even if lambda.z.time.last != tlast (#448)
+  exp_error <- "More than one half-life calculation was attempted on the following rows: 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11"
+  expect_error(suppressMessages(get_halflife_points(o_data)), regexp = exp_error)
+  expect_error(suppressMessages(get_halflife_points(o_nca)), regexp = exp_error)
+
+  ## Half-life points are selected right even if lambda.z.time.last != tlast (#448) ----
   d_conc <- data.frame(
     conc = c(1, 0.5, 0.25, 0.125, 0.07, 0, 0.01),
     time = c(0, 1, 2, 3, 4, 5, 6),
@@ -433,21 +488,25 @@ test_that("get_halflife_points", {
   o_conc <- PKNCAconc(d_conc, formula = conc ~ time | subject)
   o_data <- PKNCAdata(o_conc, intervals = data.frame(start = 0, end = Inf, half.life = TRUE))
   o_nca <- suppressMessages(pk.nca(o_data))
-  hl_points <- suppressMessages(get_halflife_points(o_nca))
+  hl_points1 <- suppressMessages(get_halflife_points(o_data))
+  hl_points2 <- suppressMessages(get_halflife_points(o_nca))
   # Note that BLQ times are included
+  expect_equal(hl_points1, hl_points2)
   expect_equal(
-    d_conc$time[which(hl_points)],
+    d_conc$time[which(hl_points2)],
     1:6,
     info = "get_halflife_points uses lambda.z.time.last, not tlast"
   )
-  # with exclusion of Tlast
+  ## with exclusion of Tlast ----
   o_conc <- PKNCAconc(d_conc, formula = conc ~ time | subject, exclude_half.life = "exclude_hl")
   o_data <- PKNCAdata(o_conc, intervals = data.frame(start = 0, end = Inf, half.life = TRUE))
   o_nca <- suppressMessages(pk.nca(o_data))
-  hl_points <- suppressMessages(get_halflife_points(o_nca))
+  hl_points1 <- suppressMessages(get_halflife_points(o_data))
+  hl_points2 <- suppressMessages(get_halflife_points(o_nca))
   # lambda.z.time.last should be 4, tlast is 6 and excluded
+  expect_equal(hl_points1, hl_points2)
   expect_equal(
-    d_conc$time[which(hl_points)],
+    d_conc$time[which(hl_points2)],
     1:4,
     info = "get_halflife_points uses lambda.z.time.last, not tlast"
   )